fix: Cleaner architecture + separate get and refresh JWT
