+++ backend/src/application/handlers/mod.rs
@@ -1,1 +1,2 @@
 pub mod items;
+pub mod users;
