+++ backend/src/application/handlers/users/mod.rs
@@ -1,0 +1,9 @@
+pub mod get_jwt;
+// pub mod log_out;
+pub mod refresh_jwt;
+pub mod register;
+
+pub use get_jwt::get_jwt;
+// pub use log_out::log_out;
+pub use refresh_jwt::refresh_jwt;
+pub use register::register;
