+++ backend/src/application/routes/mod.rs
@@ -1,4 +1,4 @@
 pub mod health_check;
 pub mod items;
 pub mod protected_endpoint;
-pub mod user;
+pub mod users;
