--- backend/src/application/routes/user/mod.rs
@@ -1,2 +1,0 @@
-pub mod get_jwt;
-pub mod register;
