+++ backend/tests/helper/mock_hasher.rs
@@ -1,4 +1,5 @@
 use ethistyle::domain::Hasher;
+use tracing_log::log::info;
 
 pub struct MockHasher;
 
@@ -6,8 +7,12 @@ impl Hasher for MockHasher {
     fn hash_password(&self, password: &str) -> Result<String, String> {
         Ok(password.to_string())
     }
-    fn verify(&self, user_password: &str, db_hash: &str) -> bool {
-        user_password == db_hash
+    fn verify(&self, password_sent: &str, db_hash: &str) -> bool {
+        info!(
+            "password_sent: {}, password in db: {}",
+            password_sent, db_hash
+        );
+        password_sent == db_hash
     }
 }
