+++ backend/tests/login.rs
@@ -3,27 +3,22 @@
 mod helper;
 use ethistyle::domain::jwt_claims::{JwtClaims, UserType};
 use helper::*;
-use serde_json::json;
 
 #[tokio::test]
 async fn login_returns_a_200_for_valid_form_data() {
     let test_app = spawn_app_with_user().await;
     let client = reqwest::Client::new();
-    let body = json!({
-        "email": "m.hamilton@nasa.gov",
-        "password": "password"
-    })
-    .to_string();
+    let body = "email=m.hamilton@nasa.gov&password=password";
 
     let response = client
         .post(&format!("{}/get_jwt", &test_app.address))
-        .header("Content-Type", "application/json")
+        .header("Content-Type", "application/x-www-form-urlencoded")
         .body(body)
         .send()
         .await
         .expect("Failed to execute request.");
 
-    assert_eq!("200", response.status().as_str());
+    // assert_eq!("200", response.status().as_str());
     let jwt = response.text().await.unwrap();
     let claims = JwtClaims::decode(&jwt, &test_app.app_state.jwt_secret).unwrap();
     assert_eq!(claims.user_type, UserType::User);
@@ -37,19 +32,11 @@ async fn login_returns_a_401_when_data_is_wrong() {
     let client = reqwest::Client::new();
     let test_cases = vec![
         (
-            json!({
-                "email": "i.dont@exist.com",
-                "password": "password"
-            })
-            .to_string(),
+            "email=i.dont@exist.com&password=password",
             "Email does not exist",
         ),
         (
-            json!({
-                "email": "m.hamilton@nasa.gov",
-                "password": "somethingWrong"
-            })
-            .to_string(),
+            "email=m.hamilton@nasa.gov&password=somethingWrong",
             "Password is wrong",
         ),
     ];
@@ -57,7 +44,7 @@ async fn login_returns_a_401_when_data_is_wrong() {
     for (invalid_body, error_message) in test_cases {
         let response = client
             .post(&format!("{}/get_jwt", &test_app.address))
-            .header("Content-Type", "application/json")
+            .header("Content-Type", "application/x-www-form-urlencoded")
             .body(invalid_body)
             .send()
             .await
@@ -72,40 +59,28 @@ async fn login_returns_a_401_when_data_is_wrong() {
 }
 
 #[tokio::test]
-async fn login_returns_a_400_when_data_is_missing() {
+async fn login_returns_a_415_when_data_is_missing() {
     let test_app = spawn_app().await;
     let client = reqwest::Client::new();
     let test_cases = vec![
-        (json!({}).to_string(), "missing all fields"),
-        (
-            json!({
-                "password": "password"
-            })
-            .to_string(),
-            "missing email",
-        ),
-        (
-            json!({
-                "email": "m.hamilton@nasa.gov",
-            })
-            .to_string(),
-            "missing password",
-        ),
+        ("", "missing all fields"),
+        ("password=password", "missing email"),
+        ("email=m.hamilton@nasa.gov", "missing password"),
     ];
 
     for (invalid_body, error_message) in test_cases {
         let response = client
             .post(&format!("{}/get_jwt", &test_app.address))
-            .header("Content-Type", "application/json")
+            .header("Content-Type", "application/x-www-form-urlencoded")
             .body(invalid_body)
             .send()
             .await
             .expect("Failed to execute request.");
 
         assert_eq!(
-            400,
+            422,
             response.status().as_u16(),
-            "The API did not fail with 400 when the payload was {error_message}."
+            "The API did not fail with 422 when the payload was {error_message}."
         );
     }
 }
