+++ backend/tests/register.rs
@@ -2,22 +2,16 @@
 
 mod helper;
 use helper::*;
-use serde_json::json;
 
 #[tokio::test]
 async fn register_returns_a_200_for_valid_form_data() {
     let test_app = spawn_app().await;
     let client = reqwest::Client::new();
-    let body = json!({
-        "name":"Margaret Hamilton",
-        "email":"m.hamilton@nasa.gov",
-        "password":"password"
-    })
-    .to_string();
+    let body = "name=Margaret Hamilton&email=m.hamilton@nasa.gov&password=password";
 
     let response = client
         .post(&format!("{}/register", &test_app.address))
-        .header("Content-Type", "application/json")
+        .header("Content-Type", "application/x-www-form-urlencoded")
         .body(body)
         .send()
         .await
@@ -45,17 +39,14 @@ async fn register_returns_a_422_when_data_is_missing() {
     let test_app = spawn_app().await;
     let client = reqwest::Client::new();
     let test_cases = vec![
-        (String::from("{}"), "missing all fields"),
+        ("", "missing all fields"),
         (
-            json!({"email": "m.hamilton@nasa.gov", "password": "password"}).to_string(),
+            "email=m.hamilton@nasa.gov&password=password",
             "missing username",
         ),
+        ("name=Margaret Hamilton&password=password", "missing email"),
         (
-            json!({"name": "Margaret Hamilton", "password": "password"}).to_string(),
-            "missing email",
-        ),
-        (
-            json!({"name": "Margaret Hamilton", "email": "m.hamilton@nasa.gov"}).to_string(),
+            "name=Margaret Hamilton&email=m.hamilton@nasa.gov",
             "missing password",
         ),
     ];
@@ -63,16 +54,16 @@ async fn register_returns_a_422_when_data_is_missing() {
     for (invalid_body, error_message) in test_cases {
         let response = client
             .post(&format!("{}/register", &test_app.address))
-            .header("Content-Type", "application/json")
+            .header("Content-Type", "application/x-www-form-urlencoded")
             .body(invalid_body)
             .send()
             .await
             .expect("Failed to execute request.");
 
         assert_eq!(
-            400,
+            422,
             response.status().as_u16(),
-            "The API did not fail with 400 Bad Request when the payload was {error_message}."
+            "The API did not fail with 422 Bad Request when the payload was {error_message}."
         );
     }
 }
@@ -82,33 +73,24 @@ async fn register_returns_a_422_when_some_fields_are_empty() {
     let test_app = spawn_app().await;
     let client = reqwest::Client::new();
     let test_cases = vec![
-        (
-            json!({"email": "", "name": "Margaret Hamilton"}).to_string(),
-            "email is empty",
-        ),
-        (
-            json!({"email": "m.hamilton@nasa.gov", "name": ""}).to_string(),
-            "name is empty",
-        ),
-        (
-            json!({"email": "", "name": ""}).to_string(),
-            "both are empty",
-        ),
+        ("email=&name=Margaret Hamilton", "email is empty"),
+        ("email=m.hamilton@nasa.gov&name=", "name is empty"),
+        ("email=&name=", "both are empty"),
     ];
 
     for (invalid_body, error_message) in test_cases {
         let response = client
             .post(&format!("{}/register", &test_app.address))
-            .header("Content-Type", "application/json")
+            .header("Content-Type", "application/x-www-form-urlencoded")
             .body(invalid_body)
             .send()
             .await
             .expect("Failed to execute request.");
 
         assert_eq!(
-            400,
+            422,
             response.status().as_u16(),
-            "The API did not fail with 400 Bad Request when the payload was {error_message}."
+            "The API did not fail with 422 Bad Request when the payload was {error_message}."
         );
     }
 }
